Upcoming Fixtures Addition
Adding the functionality of adding the upcoming games to the upcoming tab.

It will populate with all the games in the array list. Therefore, it will add all the games stored in the database.

todo: Ensure it uses a callback to ensure the data has been read in before trying to access the data.